# app/backend/swaggerdoc.json
{
    "swagger": "2.0",
    "info": {
        "description": "Manual Técnico do VolunteeringIPS",
        "version": "1.0.0",
        "title": "VolunteeringIPS Doc"
    },
    "basePath": "/api",
...
    "schemes": ["http"],
    "paths": {
        "/email/notify": {
            "post": {
                "summary": "Used to send an email to an user",
                "consumes": [
                    "application/json"
                ],
                "produces": ["application/json"],
                "tags": [
                    "Notifications"
                ],
                "parameters": [
                    {
                        "in": "body",
                        "name": "notification",
                        "description": "The notification to create",
                        "schema": {
                            "type": "object",
                            "$ref": "#/definitions/Notification"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Email successfully sent"
                    },
                    "401": {
                        "description": "Email not sent due to unauthorization (invalid token)"
                    }
                }
            }
        },
...
    },
    "definitions": {
        "Notification": {
            "type": "object",
            "required": [
                "action"
            ],
            "properties": {
                "auth": {
                    "type": "string"
                },
                "action": {
                    "type": "string"
                },
                "recipient": {
                    "type": "string"
                },
                "params": {
                    "type": "object"
                }
            }
        },
...
    }
}